Fix pandas Series to scalar conversion errors
- Add explicit float() conversions in sizing_advisor._analyze_profile
- Add explicit float() conversions in revenue_streams._analyze_profile
- Add explicit float() conversions in growth_projector._analyze_base_profile
- Fix profile_df.get() usage in stream.py (DataFrames don't have .get())

Fixes: int() argument must be a string error

+++ app/engine/sizing_advisor.py
@@ -266,12 +266,13 @@ class SizingAdvisor:
         afname_col = 'afname_kwh' if 'afname_kwh' in df.columns else 'afname'
         afname = df[afname_col].fillna(0)
 
-        peak_kw = afname.max() / interval_hours
-        avg_kw = afname.mean() / interval_hours
-        baseload_kw = afname.quantile(0.10) / interval_hours
+        # Explicitly convert to float to avoid pandas Series issues
+        peak_kw = float(afname.max()) / interval_hours
+        avg_kw = float(afname.mean()) / interval_hours
+        baseload_kw = float(afname.quantile(0.10)) / interval_hours
 
         days = len(df) / 96
-        annual_kwh = (afname.sum() / days) * 365 if days > 0 else 0
+        annual_kwh = (float(afname.sum()) / days) * 365 if days > 0 else 0
 
         return {
             "peak_kw": peak_kw,
